+++ app/Http/Controllers/Instructor/CourseController.php
@@ -218,7 +218,9 @@ class CourseController extends Controller
         $course->status = 2;
         $course->save();
 
-        $course->observation->delete();
+        if($course->observation){
+            $course->observation->delete();
+        }
 
         return redirect()->route('instructor.courses.edit', $course );
     }
